requirements.txt
matplotlib
reportlab
arabic-reshaper
python-bidi
